LPS-67168 Tests
