+++ modules/apps/collaboration/blogs/blogs-test/bnd.bnd
@@ -1,6 +1,9 @@
 Bundle-Name: Liferay Blogs Test
 Bundle-SymbolicName: com.liferay.blogs.test
 Bundle-Version: 1.0.0
+Import-Package:\
+	!com.dumbster.*,\
+	*
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title: Blogs
 -includeresource: test-classes/integration
