+++ modules/apps/collaboration/blogs/blogs-test/src/testIntegration/java/com/liferay/blogs/search/test/BlogsEntrySearchTest.java
@@ -1,0 +1,176 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.blogs.search.test;
+
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
+import com.liferay.blogs.kernel.model.BlogsEntry;
+import com.liferay.blogs.service.BlogsEntryLocalServiceUtil;
+import com.liferay.portal.kernel.model.BaseModel;
+import com.liferay.portal.kernel.service.ServiceContext;
+import com.liferay.portal.kernel.test.rule.AggregateTestRule;
+import com.liferay.portal.kernel.test.rule.Sync;
+import com.liferay.portal.kernel.test.rule.SynchronousDestinationTestRule;
+import com.liferay.portal.kernel.test.util.TestPropsValues;
+import com.liferay.portal.search.test.BaseSearchTestCase;
+import com.liferay.portal.service.test.ServiceTestUtil;
+import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
+import com.liferay.portlet.blogs.util.test.BlogsTestUtil;
+
+import org.junit.Before;
+import org.junit.ClassRule;
+import org.junit.Ignore;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+/**
+ * @author Eudaldo Alonso
+ */
+@RunWith(Arquillian.class)
+@Sync
+public class BlogsEntrySearchTest extends BaseSearchTestCase {
+
+	@ClassRule
+	@Rule
+	public static final AggregateTestRule aggregateTestRule =
+		new AggregateTestRule(
+			new LiferayIntegrationTestRule(),
+			SynchronousDestinationTestRule.INSTANCE);
+
+	@Before
+	@Override
+	public void setUp() throws Exception {
+		ServiceTestUtil.setUser(TestPropsValues.getUser());
+
+		super.setUp();
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testLocalizedSearch() throws Exception {
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testParentBaseModelUserPermissions() throws Exception {
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testSearchAttachments() throws Exception {
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testSearchByDDMStructureField() throws Exception {
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testSearchByKeywordsInsideParentBaseModel() throws Exception {
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testSearchExpireAllVersions() throws Exception {
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testSearchExpireLatestVersion() throws Exception {
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testSearchMyEntries() throws Exception {
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testSearchRecentEntries() throws Exception {
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testSearchStatus() throws Exception {
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testSearchVersions() throws Exception {
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testSearchWithinDDMStructure() throws Exception {
+	}
+
+	@Override
+	protected BaseModel<?> addBaseModelWithWorkflow(
+			BaseModel<?> parentBaseModel, boolean approved, String keywords,
+			ServiceContext serviceContext)
+		throws Exception {
+
+		return BlogsTestUtil.addEntryWithWorkflow(
+			TestPropsValues.getUserId(), keywords, approved, serviceContext);
+	}
+
+	@Override
+	protected void deleteBaseModel(long primaryKey) throws Exception {
+		BlogsEntryLocalServiceUtil.deleteEntry(primaryKey);
+	}
+
+	@Override
+	protected Class<?> getBaseModelClass() {
+		return BlogsEntry.class;
+	}
+
+	@Override
+	protected String getSearchKeywords() {
+		return "Title";
+	}
+
+	@Override
+	protected void moveBaseModelToTrash(long primaryKey) throws Exception {
+		BlogsEntryLocalServiceUtil.moveEntryToTrash(
+			TestPropsValues.getUserId(), primaryKey);
+	}
+
+	@Override
+	protected BaseModel<?> updateBaseModel(
+			BaseModel<?> baseModel, String keywords,
+			ServiceContext serviceContext)
+		throws Exception {
+
+		BlogsEntry entry = (BlogsEntry)baseModel;
+
+		return BlogsEntryLocalServiceUtil.updateEntry(
+			serviceContext.getUserId(), entry.getEntryId(), keywords,
+			entry.getContent(), serviceContext);
+	}
+
+}
