+++ modules/apps/collaboration/blogs/blogs-test/src/testIntegration/java/com/liferay/blogs/verify/test/BlogsServiceVerifyProcessTest.java
@@ -15,8 +15,8 @@
 package com.liferay.blogs.verify.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
+import com.liferay.portal.verify.VerifyProcess;
 import com.liferay.portal.verify.test.BaseVerifyProcessTestCase;
 import com.liferay.registry.Filter;
 import com.liferay.registry.Registry;
@@ -37,7 +37,7 @@ public class BlogsServiceVerifyProcessTest extends BaseVerifyProcessTestCase {
 
 	@ClassRule
 	@Rule
-	public static final AggregateTestRule aggregateTestRule =
+	public static final LiferayIntegrationTestRule liferayIntegrationTestRule =
 		new LiferayIntegrationTestRule();
 
 	@BeforeClass
