+++ modules/apps/web-experience/asset/asset-tags-test/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Asset Tags Test
+Bundle-SymbolicName: com.liferay.asset.tags.test
+Bundle-Version: 1.0.0
+-includeresource: test-classes/integration
