+++ modules/apps/web-experience/asset/asset-tags-test/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+	testIntegrationCompile group: "com.liferay", name: "com.liferay.arquillian.extension.junit.bridge", version: "1.0.5"
+	testIntegrationCompile project(":apps:collaboration:blogs:blogs-api")
+}
