+++ modules/apps/web-experience/asset/asset-test/build.gradle
@@ -11,5 +11,7 @@ dependencies {
 	testIntegrationCompile group: "com.liferay", name: "com.liferay.bookmarks.api", version: "2.0.0"
 	testIntegrationCompile group: "com.liferay", name: "com.liferay.dynamic.data.mapping.test.util", version: "2.0.0"
 	testIntegrationCompile group: "com.liferay", name: "com.liferay.journal.api", version: "2.0.0"
+	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "2.6.0"
 	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.0.0"
+	testIntegrationCompile project(":apps:collaboration:blogs:blogs-api")
 }
