+++ source-formatter.properties
@@ -125,6 +125,7 @@
     # column limit.
     #
     line.length.excludes=\
+        modules/apps/collaboration/blogs/blogs-test/src/testIntegration/java/com/liferay/blogs/subscription/test/BlogsSubscriptionRootContainerModelTest.java,\
         modules/apps/collaboration/item-selector/item-selector-editor-configuration/src/main/java/com/liferay/item/selector/editor/configuration/internal/BaseEditorConfigContributor.java@34,\
         modules/apps/collaboration/wiki/wiki-test/src/testIntegration/java/com/liferay/wiki/subscription/test/WikiSubscriptionContainerModelTest.java,\
         modules/apps/collaboration/wiki/wiki-test/src/testIntegration/java/com/liferay/wiki/trash/test/WikiPageDependentsTrashHandlerTest.java,\
@@ -159,7 +160,6 @@
         portal-impl/test/integration/com/liferay/portal/service/PortletPreferencesLocalServiceTest.java,\
         portal-impl/test/integration/com/liferay/portal/util/PortalImplCanonicalURLTest.java,\
         portal-impl/test/integration/com/liferay/portal/util/PortalImplLocalizedFriendlyURLTest.java,\
-        portal-impl/test/integration/com/liferay/portlet/blogs/subscriptions/BlogsSubscriptionRootContainerModelTest.java,\
         portal-impl/test/integration/com/liferay/portlet/documentlibrary/service/DLFileEntryExtensionTest.java@314,\
         portal-impl/test/integration/com/liferay/portlet/documentlibrary/service/DLFileEntryExtensionTest.java@336,\
         portal-impl/test/integration/com/liferay/portlet/documentlibrary/service/persistence/DLFileEntryFinderTest.java,\
